+++ src/main.rs
@@ -1,6 +1,6 @@
 use imgurian::commands::{
-    delete_image, favorite_image, generate_access_token, get_account, get_account_image, get_image,
-    list_account_images, update_image, upload_image,
+    delete_image, favorite_image, generate_access_token, get_account, get_account_image,
+    get_account_images_count, get_image, list_account_images, update_image, upload_image,
 };
 use imgurian::opt::Opt;
 use imgurian::result::Result;
@@ -15,6 +15,7 @@ async fn main() -> Result<()> {
         Opt::GenerateAccessToken { .. } => generate_access_token(opt).await?,
         Opt::GetAccount { .. } => get_account(opt).await?,
         Opt::GetAccountImage { .. } => get_account_image(opt).await?,
+        Opt::GetAccountImagesCount { .. } => get_account_images_count(opt).await?,
         Opt::GetImage { .. } => get_image(opt).await?,
         Opt::ListAccountImages { .. } => list_account_images(opt).await?,
         Opt::UpdateImage { .. } => update_image(opt).await?,